Added some new functions in ECS files and made it .hpp instead. Took reference from GAM200 only.

+++ BoofWoof/Core/ECS/pch.hpp
@@ -1,16 +1,13 @@
 #pragma once
-#include <memory>
-#include <unordered_map>
-#include <cassert>
-#include <bitset>
-#include <queue>
 #include <cstdint>
-#include <set>
+#include <unordered_set>
+#include <queue>
 #include <array>
-
-//#include "ComponentManager.h"
-//#include "EntityManager.h"
-//#include "SystemManager.h"
+#include <cassert>
+#include <unordered_map>
+#include <memory>
+#include <bitset>
+#include <iostream>
 
 // A simple type alias
 using Entity = std::uint32_t;
@@ -19,7 +16,7 @@ using Entity = std::uint32_t;
 const Entity MAX_ENTITIES = 5000;
 
 // A simple type alias
-using ComponentType = std::uint8_t;
+using ComponentType = std::uint32_t;
 
 // Used to define the size of arrays later on
 const ComponentType MAX_COMPONENTS = 32;
